requirements.txt
streamlit
pandas
matplotlib
yfinance
sqlalchemy
psycopg2-binary
groq
python-dotenv
google-generativeai
bcrypt